Fixed typecast problems and removed referenced to uci.gef.contrib from sources.
git-svn-id: http://argouml.tigris.org/svn/argouml/trunk@6 a161b567-7d1e-0410-9ef9-912c70fedb3f

+++ src/uci/gef/RedrawManager.java
@@ -119,10 +119,10 @@ public class RedrawManager implements Runnable {
 
   public static void setFramesPerSecond(float fps) {
     if (fps > 100.0 || fps < 0.5) return;
-    int _timeDelay = (int) 1000/fps;
+    int _timeDelay = (int) (1000/fps);
   }
 
-  public static float getFramesPerSecond() { return 1000.0 / _timeDelay; }
+  public static float getFramesPerSecond() { return (float) (1000.0 / _timeDelay); }
 
   ////////////////////////////////////////////////////////////////
   // managing damage
